refactor: AuctionForm의 기본 값을 가져오는 메서드 파일로 분리 (#147)
